Updated the hw for lesson 5. Moved getSessionFactory() method from AbstractEntityDao to a new util class, HibernateUtil. All works correctly.
