+++ src/main/java/ru/geekbrains/lesson5/hw/utils/HibernateUtil.java
@@ -1,0 +1,21 @@
+package ru.geekbrains.lesson5.hw.utils;
+
+import org.hibernate.SessionFactory;
+import org.hibernate.boot.Metadata;
+import org.hibernate.boot.MetadataSources;
+import org.hibernate.boot.registry.StandardServiceRegistry;
+import org.hibernate.boot.registry.StandardServiceRegistryBuilder;
+
+public class HibernateUtil {
+    private static SessionFactory sessionFactory;
+
+    public static SessionFactory getSessionFactory() {
+        if (sessionFactory == null) {
+            StandardServiceRegistry standardRegistry = new StandardServiceRegistryBuilder().configure().build();
+            Metadata metadata = new MetadataSources(standardRegistry).getMetadataBuilder().build();
+            sessionFactory = metadata.getSessionFactoryBuilder().build();
+        }
+        return sessionFactory;
+    }
+
+}
